Extract `{Assert,Refute}Match` cop implementations
These two cops have near identical implementations. Extracting them
allows us to make further changes in a single place.
