+++ lib/rubocop/cop/minitest/assert_match.rb
@@ -20,48 +20,31 @@ module RuboCop
       #
       class AssertMatch < Base
         include ArgumentRangeHelper
+        include AssertRefuteMatchHelper
         extend AutoCorrector
 
         MSG = 'Prefer using `assert_match(%<preferred>s)`.'
         RESTRICT_ON_SEND = %i[assert assert_operator].freeze
 
-        def_node_matcher :assert_match, <<~PATTERN
+        def_node_matcher :match_assertion, <<~PATTERN
           {
             (send nil? :assert (send $_ {:match :match? :=~} $_) $...)
             (send nil? :assert_operator $_ (sym :=~) $_ $...)
           }
         PATTERN
 
-        # rubocop:disable Metrics/AbcSize
         def on_send(node)
-          assert_match(node) do |expected, actual, rest_args|
-            basic_arguments = order_expected_and_actual(expected, actual)
-            preferred = (message_arg = rest_args.first) ? "#{basic_arguments}, #{message_arg.source}" : basic_arguments
-            message = format(MSG, preferred: preferred)
-
-            add_offense(node, message: message) do |corrector|
-              corrector.replace(node.loc.selector, 'assert_match')
-
-              range = if node.method?(:assert)
-                        node.first_argument
-                      else
-                        node.first_argument.source_range.begin.join(node.arguments[2].source_range.end)
-                      end
-
-              corrector.replace(range, basic_arguments)
-            end
-          end
+          check_match_assertion(node)
         end
-        # rubocop:enable Metrics/AbcSize
 
         private
 
-        def order_expected_and_actual(expected, actual)
-          if actual.regexp_type?
-            [actual, expected]
-          else
-            [expected, actual]
-          end.map(&:source).join(', ')
+        def basic_preferred_assertion_method_name
+          :assert
+        end
+
+        def preferred_assertion_method_name
+          :assert_match
         end
       end
     end
